minor fix to running parser and typechecker
