+++ GameTheoryLanguageTests/ParserTests.cs
@@ -226,19 +226,4 @@ public class ParserTests
             Console.WriteLine(parser.program().ToStringTree());
         }
     }
-    public class ErrorStrategy : DefaultErrorStrategy
-    {
-        public override void ReportError(Parser recognizer, RecognitionException e)
-        {
-            throw new ParserException();
-        }
-    }
-    public class ErrorListener : BaseErrorListener
-    {
-        public override void SyntaxError(TextWriter output, IRecognizer recognizer, IToken offendingSymbol, int line, int charPositionInLine, string msg, RecognitionException e)
-        {
-            Console.WriteLine(msg + " in line " + line + " position " + charPositionInLine);
-            throw new ParserException();
-        }
-    }
 }
